+++ drift/CHANGELOG.md
@@ -7,6 +7,7 @@
   Thanks to [@westito](https://github.com/westito).
 - Allow the generator to emit correct SQL code when using arrays with the
   `new_sql_code_generation` option in specific scenarios.
+- Add support for [strict tables](https://sqlite.org/stricttables.html) in `.drift` files.
 - Add the `generatedAs` method to declare generated columns for Dart tables.
 - Add `OrderingTerm.random` to fetch rows in a random order.
 - Improved support for pausing query stream subscriptions. Instead of buffering events,